Second checkin, First as Lithium.

+++ findrox.py
@@ -1,7 +1,7 @@
 # Most of the common code needed by ROX applications is in ROX-Lib2.
 # Except this code, which is needed to find ROX-Lib2 in the first place!
 
-# Just make sure you import findrox before importing anything inside
+# Just make sure you run findrox.version() before importing anything inside
 # ROX-Lib2...
 
 import os, sys
@@ -11,23 +11,40 @@ import string
 def version(major, minor, micro):
 	"""Find ROX-Lib2, with a version >= (major, minor, micro), and
 	add it to sys.path. If version is missing or too old, either
-	prompt the user, or (if possible) upgrade it automatically."""
+	prompt the user, or (if possible) upgrade it automatically.
+	If 'rox' is already in PYTHONPATH, just use that (assume the injector
+	is being used)."""
+	try:
+		import rox
+	except ImportError:
+		pass
+	else:
+		#print "Using ROX-Lib in PYTHONPATH"
+		if (major, minor, micro) > rox.roxlib_version:
+			print >>sys.stderr, "WARNING: ROX-Lib version " \
+				"%d.%d.%d requested, but using version " \
+				"%d.%d.%d from %s" % \
+				(major, minor, micro,
+				 rox.roxlib_version[0],
+				 rox.roxlib_version[1],
+				 rox.roxlib_version[2],
+				 rox.__file__)
+		return
 
-	if os.path.exists('/uri/0install/rox.sourceforge.net'):
+	if not os.getenv('ROXLIB_DISABLE_ZEROINSTALL') and os.path.exists('/uri/0install/rox.sourceforge.net'):
 		# We're using ZeroInstall. Good :-)
 		zpath = '/uri/0install/rox.sourceforge.net/lib/ROX-Lib2/' \
-			'latest'
-		if os.path.exists(zpath):
-			vs = os.readlink(zpath).split('-')[-1]
-			v = map(int, vs.split('.'))
-			if v[0] < major or v[1] < minor or v[2] < micro:
-				if os.system('cd /uri/0install/rox.sourceforge.net; 0refresh'):
-					report_error('Using ROX-Lib in Zero Install, but cached version (%s) is too old (need %d.%d.%d) and updating failed (is zero-install running?)' % (vs, major, minor, micro))
-			sys.path.append(zpath + '/python')
-			return
-		print >>sys.stderr, "Using Zero Install, but failed to " \
-			"fetch", zpath, "-- trying non-0install system."
-
+			'latest-2'
+		if not os.path.exists(zpath):
+			os.system('0refresh rox.sourceforge.net')
+			assert os.path.exists(zpath)
+		vs = os.readlink(zpath).split('-')[-1]
+		v = tuple(map(int, vs.split('.')))
+		if v < (major, minor, micro):
+			if os.system('0refresh rox.sourceforge.net'):
+				report_error('Using ROX-Lib in Zero Install, but cached version (%s) is too old (need %d.%d.%d) and updating failed (is zero-install running?)' % (vs, major, minor, micro))
+		sys.path.append(zpath + '/python')
+		return
 	try:
 		path = os.environ['LIBDIRPATH']
 		paths = string.split(path, ':')
@@ -40,13 +57,19 @@ def version(major, minor, micro):
 		if exists(p):
 			# TODO: check version is new enough
 			sys.path.append(os.path.join(p, 'python'))
-			break
-	else:
-		report_error("This program needs ROX-Lib2 to run.\n" + \
-			"I tried all of these places:\n\n" + \
-			string.join(paths, '\n') + '\n\n' + \
-			"ROX-Lib2 is available from:\n" + \
-			"http://rox.sourceforge.net")
+			import rox
+			if major == 1 and minor == 9 and micro < 10:
+				return	# Can't check version
+			if not hasattr(rox, 'roxlib_version'):
+				break
+			if (major, minor, micro) <= rox.roxlib_version:
+				return	# OK
+	report_error("This program needs ROX-Lib2 (version %d.%d.%d) " % \
+		(major, minor, micro) + "to run.\n" + \
+		"I tried all of these places:\n\n" + \
+		string.join(paths, '\n') + '\n\n' + \
+		"ROX-Lib2 is available from:\n" + \
+		"http://rox.sourceforge.net")
 
 def report_error(err):
 	"Write 'error' to stderr and, if possible, display a dialog box too."
